add migration and mapper

+++ WeightWatchers/WeightWatchers.Services/Models/Card.cs
@@ -9,9 +9,9 @@ namespace WeightWatchers.Services.Models
         public int id { get; set; }
         public int subscriberId { get; set; }
         public DateTime openDate { get; set; }
-        public int BMI { get; set; }
-        public int height { get; set; }
-        public int weight { get; set; }
+        public float BMI { get; set; }
+        public float height { get; set; }
+        public float weight { get; set; }
 
         public DateTime updateDate { get; set; }
